% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bollen_stine_transform}
\alias{bollen_stine_transform}
\title{Bollen-Stine transformer}
\usage{
bollen_stine_transform(object)
}
\arguments{
\item{object}{A \code{lavaan} object.}
}
\value{
A list of Bollen-Stine transformed data.
}
\description{
Bollen-Stine transformer
}
\keyword{internal}
