% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap \code{lavaan} models.}
\usage{
bootstrap(m0, m1 = NULL, data)
}
\arguments{
\item{m0, m1}{\code{lavaan} objects. Data is sample from \code{m0} and fitted with
\code{m0} and \code{m1}. If \code{m1} is \code{NULL}, only \code{m0} is fitted.}

\item{data}{The data used to sample from, e.g. Bollen-Stine transformed
data.}
}
\value{
A bootstrapped \code{lavaan} object.
}
\description{
Bootstrap \code{lavaan} models.
}
\keyword{internal}
