% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{chunk}
\alias{chunk}
\title{Split vector \code{x} into \code{n} chunks of equal size.}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{Input vector.}

\item{n}{Desired output length.}
}
\value{
List of \code{n} vectors.
}
\description{
Split vector \code{x} into \code{n} chunks of equal size.
}
\keyword{internal}
