% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{compute_pvalues}
\alias{compute_pvalues}
\title{Compute p-values for one parsed test specification.}
\usage{
compute_pvalues(
  m0,
  unbiased,
  trad,
  eba,
  peba,
  pols,
  chisq = c("ml", "rls"),
  m1 = NULL,
  A.method = "delta",
  fiml.convention = "observed"
)
}
\description{
Compute p-values for one parsed test specification.
}
\keyword{internal}
