% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{eigen_pvalues}
\alias{eigen_pvalues}
\title{Get eigenvalue-based p-values.}
\usage{
eigen_pvalues(chisq, eig, candidates)
}
\arguments{
\item{chisq}{Chi-square fit value from a lavaan object.}

\item{eig}{Eigenvalues of the UG matrix.}
}
\value{
List of eigenvalue-based p-values. \code{psb} is Satorra--Bentler,
\code{pfull} is based on every p-value, while \code{phalf} is based on the
half of the (largest) p-values.
}
\description{
Get eigenvalue-based p-values.
}
\keyword{internal}
