% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiml_fmg.R
\name{fiml_align_K0}
\alias{fiml_align_K0}
\title{Align an H0 full-parameter basis to H1's parameter ordering.}
\usage{
fiml_align_K0(m0, m1, K0)
}
\description{
Align an H0 full-parameter basis to H1's parameter ordering.
}
\keyword{internal}
