% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiml_fmg.R
\name{fiml_align_parameter_basis}
\alias{fiml_align_parameter_basis}
\title{Align an H0 parameter basis to H1's full-parameter ordering.}
\usage{
fiml_align_parameter_basis(m0, m1, basis0)
}
\description{
Align an H0 parameter basis to H1's full-parameter ordering.
}
\keyword{internal}
