% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiml_fmg.R
\name{fiml_check_same_data}
\alias{fiml_check_same_data}
\title{Verify that two FIML fits use identical observations.}
\usage{
fiml_check_same_data(m0, m1)
}
\description{
Verify that two FIML fits use identical observations.
}
\keyword{internal}
