% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiml_fmg.R
\name{fiml_K_matrix}
\alias{fiml_K_matrix}
\title{A basis mapping effective parameters to lavaan's full free parameters.}
\usage{
fiml_K_matrix(fit)
}
\description{
This handles lavaan's explicit equality-constraint basis, simple equality
constraints, and equality constraints combined with inequalities or bounds.
}
\keyword{internal}
