% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiml_fmg.R
\name{fiml_nested_ingredients}
\alias{fiml_nested_ingredients}
\title{Gamma-free nested FIML ingredients in H1's effective parameter space.}
\usage{
fiml_nested_ingredients(m1, fiml.convention = c("observed", "lavaan"))
}
\description{
Gamma-free nested FIML ingredients in H1's effective parameter space.
}
\keyword{internal}
