% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiml_fmg.R
\name{fiml_one_group_matrix}
\alias{fiml_one_group_matrix}
\title{Extract the single matrix returned by a one-group lavaan inspection.}
\usage{
fiml_one_group_matrix(x, what)
}
\description{
Extract the single matrix returned by a one-group lavaan inspection.
}
\keyword{internal}
