% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiml_fmg.R
\name{fiml_saturated_moments}
\alias{fiml_saturated_moments}
\title{Saturated observed-data FIML eta-space pieces.}
\usage{
fiml_saturated_moments(fit)
}
\description{
Saturated observed-data FIML eta-space pieces.
}
\keyword{internal}
