% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fit_df}
\alias{fit_df}
\title{Extract model degrees of freedom through lavaan's public API.}
\usage{
fit_df(fit)
}
\description{
Extract model degrees of freedom through lavaan's public API.
}
\keyword{internal}
