% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_ugamma.R
\name{gamma}
\alias{gamma}
\title{Get gamma from a model.}
\usage{
gamma(m1, unbiased = 1, m0 = NULL)
}
\arguments{
\item{m1}{Model to extract gamma from.}

\item{unbiased}{Biased (1), unbiased (2), or both (3).}

\item{m0}{Optional second model, used if \code{m1} does not work.}
}
\value{
List of (un)biased gammas.
}
\description{
Get gamma from a model.
}
\keyword{internal}
