% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma_est_adf}
\alias{gamma_est_adf}
\title{Asymptotically distribution free covariance matrix.}
\usage{
gamma_est_adf(x)
}
\arguments{
\item{x}{Data.}
}
\value{
Estimate of the ADF covariance matrix.
}
\description{
Asymptotically distribution free covariance matrix.
}
\keyword{internal}
