% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma_est_nt}
\alias{gamma_est_nt}
\title{Normal theory gamma matrix}
\usage{
gamma_est_nt(sigma)
}
\arguments{
\item{sigma}{Covariance matrix of the data.}
}
\value{
Normal theory gamma matrix.
}
\description{
Code obtained from \code{lavaan}:
https://github.com/yrosseel/lavaan/blob/6f047c800206d23f246d484b9522295257614222/R/lav_matrix.R
}
\details{
Calculate the gamma matrix from a matrix of observations.
}
\keyword{internal}
