% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma_est_unbiased}
\alias{gamma_est_unbiased}
\title{Unbiased asymptotic covariance matrix.}
\usage{
gamma_est_unbiased(
  x,
  n = NULL,
  sigma = NULL,
  gamma_adf = NULL,
  gamma_nt = NULL
)
}
\arguments{
\item{x}{Data.}

\item{sigma}{Unbiased covariance matrix of the data.}

\item{gamma_adf}{The \code{gamma} matrix. If \code{NULL}, computes gamma from \code{x} and \code{sigma}.}

\item{gamma_nt}{The \code{gamma} matrix under normal theory.
If \code{NULL}, computes \code{gamma_nt} from \code{sigma}.}
}
\value{
Unbiased asymptotic covariance matrix.
}
\description{
Unbiased asymptotic covariance matrix.
}
\keyword{internal}
