% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma_to_gamma_unbiased}
\alias{gamma_to_gamma_unbiased}
\title{Calculate unbiase gamma from gamma and object.}
\usage{
gamma_to_gamma_unbiased(gammas, object)
}
\arguments{
\item{gammas}{List of gammas for each group.}

\item{object}{\code{lavaan} object that corresponds to gamma.}
}
\value{
List of unbiased gammas.
}
\description{
Calculate unbiase gamma from gamma and object.
}
\keyword{internal}
