% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma_unbiased}
\alias{gamma_unbiased}
\title{Calculate unbiased gamma from lavaan object and a gamma matrix.}
\usage{
gamma_unbiased(obj, gamma)
}
\arguments{
\item{obj, gamma}{Object and gamma matrix.}
}
\value{
Unbiased gamma.
}
\description{
WORKS ONLY FOR MODELS WITH NO MEAN STRUCTURE.
}
\keyword{internal}
