% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_saturated}
\alias{get_saturated}
\title{Calculate a saturated model.}
\usage{
get_saturated(object)
}
\arguments{
\item{object}{A \code{lavaan} object.}
}
\value{
A fitted saturated model.
}
\description{
Calculate a saturated model.
}
\keyword{internal}
