% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{laavan_tests}
\alias{laavan_tests}
\alias{scaled_and_shifted}
\title{Calculate the scaled and shifted / the mean-variance adjusted p-value}
\usage{
scaled_and_shifted(chisq, lambdas)
}
\arguments{
\item{chisq}{Chi-square fit value from a lavaan object.}

\item{lambdas}{Eigenvalues of UG matrix.}
}
\value{
The scaled and shifted p-value or the mean-variance adjusted p-value.
}
\description{
Calculate the scaled and shifted / the mean-variance adjusted p-value
}
\keyword{internal}
