% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{lav_ugamma_nested_2000}
\alias{lav_ugamma_nested_2000}
\title{Calculate nested ugamma.}
\usage{
lav_ugamma_nested_2000(m0, m1, gamma, a = NULL)
}
\arguments{
\item{m0, m1}{Two nested \code{lavaan} objects.}

\item{gamma}{Gamma weighted by groups.}

\item{a}{The \code{A} restriction matrix. If \code{NULL}, it is calculated from the
two models' public inspected delta matrices.}
}
\value{
Ugamma for nested object.
}
\description{
This can also be used with restrictions.
}
\keyword{internal}
