% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{lavaan_lambdas}
\alias{lavaan_lambdas}
\title{Biased single-model spectrum exposed by lavaan.}
\usage{
lavaan_lambdas(object, df)
}
\description{
Biased single-model spectrum exposed by lavaan.
}
\keyword{internal}
