% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{lower_vec_indices}
\alias{lower_vec_indices}
\title{Obtain indices of lower or upper triangular matrix in vec indices.}
\usage{
lower_vec_indices(n = 1L, diagonal = TRUE)
}
\arguments{
\item{n}{Dimension of square matrix.}

\item{diagonal}{If \code{TRUE}, includes the diagonal elements.}
}
\value{
Indices \code{x} so that \code{a[x] = c(a)[x]} returns the elements
of the lower (upper) diagonal matrix in row-wise (column-wise)
order.
}
\description{
Code obtained from \code{lavaan}:
https://github.com/yrosseel/lavaan/blob/6f047c800206d23f246d484b9522295257614222/R/lav_matrix.R
}
\keyword{internal}
