% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helper.R
\name{model_delta_matrix}
\alias{model_delta_matrix}
\title{Stack lavaan's per-group moment derivatives.}
\usage{
model_delta_matrix(fit)
}
\description{
Stack lavaan's per-group moment derivatives.
}
\keyword{internal}
