% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helper.R
\name{model_effective_delta}
\alias{model_effective_delta}
\title{Moment derivatives with respect to effective model parameters.}
\usage{
model_effective_delta(fit)
}
\description{
Moment derivatives with respect to effective model parameters.
}
\keyword{internal}
