% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nanull}
\alias{nanull}
\title{Return non-missing values and convert \code{NA} to \code{NULL}.}
\usage{
nanull(x)
}
\description{
Return non-missing values and convert \code{NA} to \code{NULL}.
}
\keyword{internal}
