% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{pvalue_internal}
\alias{pvalue_internal}
\alias{pvalues_one}
\title{P value function for one and two arguments.}
\usage{
pvalues_one(
  object,
  unbiased,
  trad,
  eba,
  peba,
  pols,
  chisq = c("trad", "rls"),
  extras = FALSE
)
}
\value{
pvalues.
}
\description{
P value function for one and two arguments.
}
\keyword{internal}
