% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{refitted_pvalue}
\alias{refitted_pvalue}
\title{Refitted p-values for a lavaan object.}
\usage{
refitted_pvalue(object, pvalue = c("pvalue.scaled", "pvalue"), ...)
}
\arguments{
\item{object}{A fitted \code{lavaan} object.}

\item{pvalue}{Passed to \code{lavaan::fitmeasures} as \code{fit.measures}.
Defaults to \code{pvalue.scaled}.}

\item{...}{Passed to \code{lavaan::sem}.}
}
\description{
Refitted p-values for a lavaan object.
}
