% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{requested_estimator}
\alias{requested_estimator}
\title{Requested lavaan estimator, before shortcut normalization.}
\usage{
requested_estimator(fit)
}
\description{
Requested lavaan estimator, before shortcut normalization.
}
\keyword{internal}
