% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{s_and_s_inv}
\alias{s_and_s_inv}
\title{Calculate s and s_inv for all subgroups of a \code{lavaan} object.}
\usage{
s_and_s_inv(object)
}
\arguments{
\item{object}{A \code{lavaan} object.}
}
\value{
A list containing s and s_inv for all subgroups of a \code{lavaan} object.
}
\description{
Calculate s and s_inv for all subgroups of a \code{lavaan} object.
}
\keyword{internal}
