% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{scaled_f}
\alias{scaled_f}
\title{Calculate the scaled_f p-value.}
\usage{
scaled_f(chisq, eig)
}
\arguments{
\item{chisq}{Chi-square fit value from a lavaan object.}

\item{eig}{eig of UG matrix.}
}
\value{
scaled f p-value.
}
\description{
Calculate the scaled_f p-value.
}
\keyword{internal}
