% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semselector.R
\name{semselector}
\alias{semselector}
\title{Select p-value for a lavaan model using the bootstrap selector.}
\usage{
semselector(
  m0,
  m1 = NULL,
  n_reps = 1000,
  distances = c("kolmogorov-smirnov", "anderson-darling", "cramer-von mises",
    "kullback-leibler", "0.05-distance")
)
}
\arguments{
\item{m0, m1}{One or two \code{lavaan} objects. If two, the first object should be
with restrictions and the second without.}

\item{n_reps}{Number of bootstrap repetitions.}

\item{distances}{A vector of strings containing the distances to calculate.
Passed to \code{distance}.}
}
\value{
An object of class \code{semselector}.
}
\description{
Select p-value for a lavaan model using the bootstrap selector.
}
