% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{semtests_warn}
\alias{semtests_warn}
\title{Signal a typed semTests warning.}
\usage{
semtests_warn(message, subclass = "semTests_warning")
}
\description{
Signal a typed semTests warning.
}
\keyword{internal}
