% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sparsify}
\alias{sparsify}
\title{Create sparse matrix}
\usage{
sparsify(mat, lim = 1e-09)
}
\arguments{
\item{mat}{Matrix input.}

\item{lim}{Elements with absolute value less than \code{lim} get set to \code{0}.}
}
\value{
Object of \code{dgCMatrix}.
}
\description{
Create sparse matrix
}
