% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{trad_pvalue}
\alias{trad_pvalue}
\title{Calculate traditional pvalues.}
\usage{
trad_pvalue(df, chisq, lambdas, type = c("pstd", "psf", "pss", "psb", "pfull"))
}
\arguments{
\item{df, chisq, lambdas, type}{Parameters needed to calculate the p-values.}
}
\value{
Traditional p-values.
}
\description{
Calculate traditional pvalues.
}
\keyword{internal}
