% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{ugamma_no_groups}
\alias{ugamma_no_groups}
\title{Calculate non-nested gamma without mean structure}
\usage{
ugamma_no_groups(object, unbiased = 1)
}
\description{
Calculate non-nested gamma without mean structure
}
\keyword{internal}
