% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{ugamma_non_nested}
\alias{ugamma_non_nested}
\title{Calculate non-nested ugamma for multiple groups.}
\usage{
ugamma_non_nested(object, unbiased = FALSE)
}
\arguments{
\item{object}{A \code{lavaan} object.}

\item{unbiased}{If \code{TRUE}, uses the unbiased gamma estimate.}
}
\value{
Ugamma for non-nested object.
}
\description{
Calculate non-nested ugamma for multiple groups.
}
\keyword{internal}
