% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{validate_tests}
\alias{validate_tests}
\title{Validate public test specifications.}
\usage{
validate_tests(tests)
}
\arguments{
\item{tests}{Character vector supplied to the public \code{tests} argument.}
}
\value{
\code{tests}, invisibly.
}
\description{
Validate public test specifications.
}
\keyword{internal}
