% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bollen_stine_transform}
\alias{bollen_stine_transform}
\title{Bollen-Stine transformer}
\usage{
bollen_stine_transform(...)
}
\arguments{
\item{...}{Models to transform.}
}
\value{
A list of Bollen-Stine transformed data.
}
\description{
Does the Bollen-Stine transform on a list of models using the implied
covariance structure of the first model.
}
\keyword{internal}
