% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap a single model with groups one time.}
\usage{
bootstrap(object, data)
}
\arguments{
\item{object}{A \code{lavaan} object.}

\item{data}{The data used to sample from, e.g. Bollen-Stein transformed
data.}
}
\value{
A bootstrapped \code{lavaan} object.
}
\description{
Bootstrap a single model with groups one time.
}
