% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Bootstrap p-values for a \code{lavaan} object.}
\usage{
bootstrapper(object, n_reps = 10)
}
\arguments{
\item{object}{A \code{lavaan} object.}

\item{n_reps}{Number of bootstrap repetitions.}
}
\value{
Bootstrapped p-values.
}
\description{
Bootstrap p-values for a \code{lavaan} object.
}
