% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Bootstrap p-values for a \code{lavaan} object.}
\usage{
bootstrapper(..., functional, n_reps = 10)
}
\arguments{
\item{...}{Several \code{lavaan} objects.}

\item{functional}{The functional to calculate. Takes a list of \code{lavaan}
objects as its argument.}

\item{n_reps}{Number of bootstrap repetitions.}
}
\value{
Bootstrapped p-values.
}
\description{
TODO: Fix bootstrap for groups.
}
\details{
TODO: Bollen-Stine
}
\keyword{internal}
