% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{distance}
\alias{distance}
\title{Calculate distances between an vector and the uniform distribution.}
\usage{
distance(x, dist = c("kolmogorov-smirnov", "anderson-darling"))
}
\arguments{
\item{x}{Vector of observations in \verb{[0,1]}.}

\item{dist}{Distance measure.}
}
\value{
The calculated distance.
}
\description{
Calculate distances between an vector and the uniform distribution.
}
