% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiml_fmg.R
\name{fiml_lambdas}
\alias{fiml_lambdas}
\title{FIML goodness-of-fit eigenvalues in saturated eta-space.}
\usage{
fiml_lambdas(fit, df)
}
\description{
FIML goodness-of-fit eigenvalues in saturated eta-space.
}
\keyword{internal}
