% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{laavan_tests}
\alias{laavan_tests}
\alias{scaled_and_shifted}
\alias{mean_var_adjusted}
\title{Calculate the scaled and shifted / the mean-variance adjusted p-value}
\usage{
scaled_and_shifted(object)

mean_var_adjusted(object)
}
\arguments{
\item{object}{\code{lavaan} object.}
}
\value{
The scaled and shifted p-value or the mean-variance adjusted p-value.
}
\description{
Calculate the scaled and shifted / the mean-variance adjusted p-value
}
