% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{laavan_tests}
\alias{laavan_tests}
\alias{scaled_and_shifted}
\title{Calculate the scaled and shifted / the mean-variance adjusted p-value}
\usage{
scaled_and_shifted(m0, m1 = NULL)
}
\arguments{
\item{m0, m1}{\code{lavaan} objects.}
}
\value{
The scaled and shifted p-value or the mean-variance adjusted p-value.
}
\description{
Calculate the scaled and shifted / the mean-variance adjusted p-value
}
