% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semselector.R
\name{plot.semselector}
\alias{plot.semselector}
\title{Plotting generic for \code{semselector} objects.}
\usage{
\method{plot}{semselector}(x, y, nrow = 3, binwidth = 0.05, ...)
}
\arguments{
\item{x}{\code{semselector} object.}

\item{y}{Ignored.}

\item{nrow}{Passed to \code{ggplot2::facet_wrap}.}

\item{binwidth}{Passed to \code{ggplot2::geom_histogram}. Defaults to \code{0.05}, to
make it easier to see the significance cutoff.}

\item{...}{Passed to \code{ggplot2::geom_histogram}.}
}
\description{
Plotting generic for \code{semselector} objects.
}
