% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{pvalue_internal}
\alias{pvalue_internal}
\alias{pvalues_one}
\alias{pvalues_two}
\title{P value function for one and two arguments.}
\usage{
pvalues_one(object)

pvalues_two(m0, m1)
}
\arguments{
\item{object, m0, m1}{\code{lavaan} object.}
}
\value{
pvalues.
}
\description{
P value function for one and two arguments.
}
\keyword{internal}
