% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{pvalue_internal}
\alias{pvalue_internal}
\alias{pvalues_one}
\alias{pvalues_two}
\title{P value function for one and two arguments.}
\usage{
pvalues_one(object, unbiased, trad, eba)

pvalues_two(m0, m1, unbiased = FALSE, trad, eba)
}
\arguments{
\item{object, m0, m1}{\code{lavaan} objects.}

\item{trad}{List of traditional p-values to calculate.}

\item{eba}{List of which eba p-values to calculate.}
}
\value{
pvalues.
}
\description{
P value function for one and two arguments.
}
\keyword{internal}
