% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{pvalues}
\alias{pvalues}
\title{Calculate p-values for a lavaan object.}
\usage{
pvalues(object)
}
\arguments{
\item{object}{A \code{lavaan} object.}
}
\value{
A named vector containing the p-values \code{pml}. \code{psb}, \code{pfull},
\code{phalf}, \code{pcf}, \code{pss}.
}
\description{
Calculate p-values for a lavaan object.
}
