% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semselector.R
\name{semselector}
\alias{semselector}
\title{Select \emph{p}-value for a lavaan model using the bootstrap selector.}
\usage{
semselector(
  m0,
  m1 = NULL,
  n_reps = 1000,
  distances = c("kolmogorov-smirnov", "anderson-darling", "cramer-von mises",
    "kullback-leibler", "0.05-distance"),
  skip_warning = FALSE
)
}
\arguments{
\item{m0, m1}{One or two \code{lavaan} objects. If two, the first object should be
with restrictions and the second without.}

\item{n_reps}{Number of bootstrap repetitions.}

\item{distances}{A vector of strings containing the distances to calculate.
Passed to \code{distance}.}

\item{skip_warning}{If \code{TRUE}, ignores bootstrapped estimates with
warnings.}
}
\value{
An object of class \code{semselector}, inheriting from \code{data.frame},
containing the winning \emph{p}-values for the selected distances.
The attributes are:
\itemize{
\item \code{boots}: The data frame of bootstrap samples.
\item \code{n_reps}: Argument passed to \code{n_reps}.
\item \code{pvalues}: The calculated p-values.
\item \code{distances} The estimated distances.
\item \code{bollen-stine}: The Bollen-Stine \emph{p}-value.
}
}
\description{
Select \emph{p}-value for a lavaan model using the bootstrap selector.
}
