% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semselector.R
\name{semselector}
\alias{semselector}
\title{Select p-value for a lavaan model using the bootstrap selector.}
\usage{
semselector(
  object,
  n_reps = 1000,
  distances = c("kolmogorov-smirnov", "anderson-darling", "cramer-von mises",
    "kullback-leibler", "0.05-distance"),
  custom = NULL,
  type = NULL
)
}
\arguments{
\item{object}{A fitted \code{lavaan} object.}

\item{n_reps}{Number of bootstrap repetitions.}

\item{distances}{A vector of strings containing the distances to calculate.
Passed to \code{distance}. \strong{Note:} \code{kullback-leibler} is very slow, hence not
in the default arguments.}

\item{custom}{List of named custom functions passed that take
\code{lavaan} objects as input.}

\item{type}{Normal parametric bootstrap / transformed samples bootstrap.}
}
\value{
An object of class \code{semselector}.
}
\description{
Select p-value for a lavaan model using the bootstrap selector.
}
