% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ugamma_nested}
\alias{ugamma_nested}
\title{Calculate nested ugamma.}
\usage{
ugamma_nested(m0, m1, a = NULL, method = "delta")
}
\arguments{
\item{m0, m1}{Two nested \code{lavaan} objects.}

\item{a}{The \code{A} matrix. If if \code{NULL}, gets calculated by
\code{lavaan:::lav_test_diff_A} with \code{method = method}.}

\item{method}{Method passed to \code{lavaan:::lav_test_diff_A}.}
}
\value{
Ugamma for non-nested object.
}
\description{
This can also be used with restrictions.
}
\keyword{internal}
