% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ugamma_non_nested}
\alias{ugamma_non_nested}
\title{Calculate non-nested ugamma for multiple groups.}
\usage{
ugamma_non_nested(object)
}
\arguments{
\item{object}{A \code{lavaan} object.}
}
\value{
Ugamma for non-nested object.
}
\description{
Calculate non-nested ugamma for multiple groups.
}
\keyword{internal}
